requirements.txt
flask
Flask-sqlalchemy
flask-login
python-docx